pyproject.toml
[tool.poetry]
name = "otel-backend"
version = "0.1.0"
description = ""
authors = ["Matthias Osswald <matthias@busykoala.io>"]
readme = "README.md"
packages = [{include = "otel_backend"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
opentelemetry-proto = "^1.21.0"
pandas = "^2.2.2"
scikit-learn = "^1.4.2"
torch = { url = "https://download.pytorch.org/whl/cpu/torch-2.3.0%2Bcpu-cp311-cp311-linux_x86_64.whl" }
torch-geometric = "^2.5.3"
matplotlib = "^3.8.4"
scikit-optimize = "^0.10.1"
seaborn = "^0.13.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"
ruff = "^0.3.5"
matplotlib = "^3.8.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 20
